Add a possibility to connect to other hosts than local
Nowadays sphinxql can connect to hosts specified in connection_params in
the configuration. We can connect to remote sphinx servers. With this
the necessity for generating sphinx.conf during startup is void.
Therefore this functionality is refactored out and a command to do this
is added (generate_shpinx_conf). Also the directories required to place
the file are created as well.